Rework non basket journey delivery details page renders to reduce duplication
