+++ test/controller/certificates/delivery.details.controller.unit.test.ts
@@ -1,9 +1,8 @@
 import chai from "chai";
 
 import { CertificateItem } from "@companieshouse/api-sdk-node/dist/services/order/certificates/types";
-import { setBackLink } from "../../../src/controllers/certificates/delivery.details.controller";
+import { setBackLink } from "../../../src/utils/delivery.details.utils";
 import { mockDissolvedCertificateItem } from "../../__mocks__/certificates.mocks";
-import { dataEmpty, fullPageFalse, fullPageTrue } from "../../__mocks__/session.mocks";
 
 describe("delivery.details.controller.unit", () => {
     describe("setBackUrl for certificate", () => {
@@ -14,7 +13,7 @@ describe("delivery.details.controller.unit", () => {
                     surname: "smith"
                 }
             } as CertificateItem;
-            chai.expect(setBackLink(certificateItem, dataEmpty)).to.equal("delivery-options");
+            chai.expect(setBackLink(certificateItem)).to.equal("delivery-options");
         });
         it("the back button link should take the user to the email options page", () => {
             const certificateItem = {
@@ -22,7 +21,7 @@ describe("delivery.details.controller.unit", () => {
                     deliveryTimescale: "same-day"
                 }
             } as CertificateItem;
-            chai.expect(setBackLink(certificateItem, dataEmpty)).to.equal("email-options");
+            chai.expect(setBackLink(certificateItem)).to.equal("email-options");
         });
     });
 
@@ -30,14 +29,14 @@ describe("delivery.details.controller.unit", () => {
         const certificateItem = mockDissolvedCertificateItem as CertificateItem;
 
         it("the back button link should take the user to the email options page for same-day delivery option", () => {
-            chai.expect(setBackLink(certificateItem, dataEmpty)).to.equal("email-options");
+            chai.expect(setBackLink(certificateItem)).to.equal("email-options");
         });
 
         it("the back button link should take the user to the delivery options page for standard delivery option", () => {
             mockDissolvedCertificateItem.itemOptions.deliveryTimescale = "standard";
             mockDissolvedCertificateItem.itemOptions.includeEmailCopy = false;
 
-            chai.expect(setBackLink(certificateItem, dataEmpty)).to.equal("delivery-options");
+            chai.expect(setBackLink(certificateItem)).to.equal("delivery-options");
         });
     });
 });
